Your commit message

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template,redirect, request
+from flask import Flask, render_template,redirect, request, flash, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
 
@@ -6,6 +6,7 @@ from datetime import datetime
 app = Flask(__name__)
 # Configure the SQLite database, relative to the app instance folder
 app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///todo.db"
+app.secret_key = "Hey this is my secret key"
 # Initialize the database object
 db = SQLAlchemy(app)
 
@@ -25,15 +26,17 @@ app.app_context().push()
 # Define the home route
 @app.route('/', methods =['GET', 'POST'])
 def home():
+    title = None
     if(request.method=='POST'):
         title = request.form['title']
         description = request.form['description']
         todo = Todo(title=title, description=description)
         db.session.add(todo)
         db.session.commit()
+        flash("Added!")
     allTodo = Todo.query.all()
     print(allTodo)
-    return render_template('index.html', allTodo = allTodo)
+    return render_template('index.html', allTodo = allTodo, todoname = title)
 
 # Define the about route
 @app.route('/update/<int:sno>', methods=['GET', 'POST'])
@@ -52,12 +55,20 @@ def update(sno):
     return render_template('update.html', todo = todo)
 
 
-@app.route('/delete/<int:sno>')
-def delete(sno):
-    todo = Todo.query.filter_by(sno=sno).first()
-    db.session.delete(todo)
-    db.session.commit()
-    return redirect('/')   
+@app.route('/delete', methods=['POST'])
+
+def delete():
+     todo_id = request.form['id']
+     todo = Todo.query.filter_by(sno=todo_id).first()
+     if todo:
+            db.session.delete(todo)
+            db.session.commit()
+            return jsonify({'success': True, 'message': 'Todo deleted successfully!'})
+     else:
+        return jsonify({'success': False, 'message': 'Todo not found!'})
+        
+ 
+    
 
 # Run the app
 if __name__ == '__main__':
